% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\name{combinations_count_fake_any_fair}
\alias{combinations_count_fake_any_fair}
\alias{combinations_count_fake_gt_fair}
\alias{combinations_count_fake_eq_fair}
\alias{combinations_count_fake_lt_fair}
\title{Return possible combinations of heads from the fake and fair coins after \code{N}
trials}
\usage{
combinations_count_fake_any_fair(comparator = function(x, y) rep(TRUE,
  length(x)), N)

combinations_count_fake_gt_fair(...)

combinations_count_fake_eq_fair(...)

combinations_count_fake_lt_fair(...)
}
\arguments{
\item{comparator}{Function to be applied for filtering of all combinations.
E.g. with \code{\link[magrittr:is_greater_than]{magrittr::is_greater_than()}}, the function will just return the
combinations where the fake coin count is higher than the fair coun count.}
}
\description{
Note that \code{combinations_count_fake_any_fair()}'s only purpose is to be used
when creating the other functions in order to avoid code duplication.
}
\section{Functions}{
\itemize{
\item \code{combinations_count_fake_gt_fair}: List combinations where the
count for the fake coin is greater than for the fair count.

\item \code{combinations_count_fake_eq_fair}: List combinations where the
count for the fake coin is equal to the fair count.

\item \code{combinations_count_fake_lt_fair}: List combinations where the
count for the fake coin is smaller than for the fair count.
}}

