% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{p_correctly_identify_coins}
\alias{p_correctly_identify_coins}
\title{How likely are we to identify the fake count after \code{N} trials?}
\usage{
p_correctly_identify_coins(N, allow_guessing = TRUE)
}
\arguments{
\item{N}{The number of tirals.}

\item{allow_guessing}{Boolean to indicate if for the case of identical
head counts, we should make a guess.}
}
\value{
A probability.
}
\description{
How likely are we to identify the fake count after \code{N} trials?
}
