% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{p_fake}
\alias{p_fake}
\title{How likely are \code{n} heads after \code{N} trials with the fake coin?}
\usage{
p_fake(n, N)
}
\arguments{
\item{n}{The number of heads.}

\item{N}{The number of tirals.}
}
\value{
A probability.
}
\description{
How likely are \code{n} heads after \code{N} trials with the fake coin?
}
