% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{p_fake_any_fair}
\alias{p_fake_any_fair}
\alias{p_fake_gt_fair}
\alias{p_fake_eq_fair}
\alias{p_fake_lt_fair}
\title{How likely is it that we have more, less or equally many fake and fair heads
after \code{N} trials?}
\usage{
p_fake_any_fair(comparator, N)

p_fake_gt_fair(...)

p_fake_eq_fair(...)

p_fake_lt_fair(...)
}
\arguments{
\item{comparator}{A function that takes two arguments and returns a boolean
value for each of them.}

\item{N}{The number of tirals.}
}
\value{
A probability.
}
\description{
\code{p_fake_any_fair()}'s only purpose is to use it for creating the other
functions.
}
\section{Functions}{
\itemize{
\item \code{p_fake_gt_fair}: Returns the probability that we have a higher
head count for the fake than for the fair coun after \code{N} trials.

\item \code{p_fake_eq_fair}: Returns the probability that we have the same
head count for the fake and for the fair coun after \code{N} trials.

\item \code{p_fake_lt_fair}: Returns the probability that we have a lower
head count for the fake than for the fair coun after \code{N} trials.
}}

