% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{p_fair}
\alias{p_fair}
\title{How likely are `n` heads after `N` trials with the fair coin?}
\usage{
p_fair(n, N)
}
\arguments{
\item{n}{The number of heads.}

\item{N}{The number of tirals.}
}
\value{
A probability.
}
\description{
How likely are `n` heads after `N` trials with the fair coin?
}
